% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{angmom_galaxy}
\alias{angmom_galaxy}
\title{Computing the angular momentum of the particles.}
\usage{
angmom_galaxy(part_data)
}
\arguments{
\item{part_data}{The particles from which you would like to calculate the
angular momentum vector, J.}
}
\value{
The numeric vector describing the magnitudes of each component of
the angular momentum.
}
\description{
The purpose of this function is to compute the angular momentum vector.
}
\examples{
  galaxy_data = data.frame("ID"=1:100, "x"=stats::rnorm(100),
                           "y"=stats::rnorm(100), "z"=stats::rnorm(100),
                           "vx"=stats::rnorm(100), "vy"=stats::rnorm(100),
                           "vz"=stats::rnorm(100), "Mass"=rep(1,100))
  J = angmom_galaxy(galaxy_data)
}
