% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC03lr}
\alias{BC03lr}
\title{BC03lr - Bruzual & Charlot 2003 Low-Resolution Spectral Templates for Stellar Populations}
\format{
A list with 4 elements:
\describe{
\item{Z}{A numeric array that describes the 6 possible metallicity bins provided by the template spectra.}
\item{Age}{A numeric array that describes the 221 possible age bins provided by the template spectra.}
\item{Wave}{A numeric array that describes the 1221 wavelengths (in Angstrom) at which the spectrum is provided.}
\item{Zspec}{A list containing the template spectrum for a galaxy. The first level of the list corresponds to
the metallicity of the population. Each list element then contains a numeric matrix, with each row describing the
spectrum that would be associated with a population of a given age.}
}
}
\source{
\url{http://www.bruzual.org/bc03}
}
\usage{
BC03lr
}
\description{
A dataset containing the grid of spectral templates for a stellar population
with a given stellar age and metallicity. At each age/metallicity bin, a
template spectrum is given in units of solar luminoisities per angstrom for
1 solar mass of star formation. Values are given at a series of wavelengths,
contained within the dataset. Note that while the first age bin is a
template for Age = 0 stars, we have re-labelled this bin with a non-zero small
number to avoid log errors.
}
\keyword{datasets}
