% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur_datacube.R
\name{blur_datacube}
\alias{blur_datacube}
\title{A function for applying seeing conditions to the mock cube}
\usage{
blur_datacube(datacube_output)
}
\arguments{
\item{datacube_output}{The list output from \code{\link{build_datacube}}.}
}
\value{
Returns a new list that contains the spectral cube produced by
\code{\link{build_datacube}}, but with the seeing conditions applied.
}
\description{
The purpose of this function is to apply a convolution kernel to each spatial
plane of the cube produced in \code{\link{build_datacube}}.
}
\examples{
ss_gadget = system.file("extdata", "SimSpin_example_Gadget_spectra.Rdata",
                        package = "SimSpin")
cube = build_datacube(simspin_file = ss_gadget,
                     telescope = telescope(type="SAMI"),
                     objective = objective())
blurred_cube = blur_datacube(datacube_output = cube)

}
