% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\docType{class}
\name{Distance-class}
\alias{Distance-class}
\alias{kpc_per_arcsec}
\alias{Mpc}
\alias{z}
\title{An S4 class to represent projected distance to a galaxy}
\usage{
kpc_per_arcsec(x)

Mpc(x)

z(x)
}
\arguments{
\item{x}{Distance object}
}
\description{
An S4 class to represent projected distance to a galaxy
}
\section{Functions}{
\itemize{
\item \code{kpc_per_arcsec()}: Query the kpc_per_arcsec of a Distance object

\item \code{Mpc()}: Query the Mpc of a Distance object

\item \code{z()}: Query the z of a Distance object

}}
\section{Slots}{

\describe{
\item{\code{z}}{Describes the projected distance to a galaxy in units of redshift.}

\item{\code{Mpc}}{Describes the projected distance to a galaxy in physical units of
Mega-parsecs.}

\item{\code{kpc_per_arcsec}}{Describes the projected distance to a galaxy in angular
units of kilo-parsecs per arcsecond.}

\item{\code{ref}}{Describes the cosmological parameters which which to compute
distances.}
}}

