% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filt_i_SDSS}
\alias{filt_i_SDSS}
\title{i-band SDSS filter responses at a range of wavelengths}
\format{
A data table with 2 variables and 89 observations:
\describe{
\item{wave}{A numeric representing the wavelength at which the relative response is recorded, given in Angstroms}
\item{response}{A numeric representing the relative response of the filter at that wavelength}
}
}
\source{
\url{https://classic.sdss.org/dr7/instruments/imager/filters/i.dat}
}
\usage{
filt_i_SDSS
}
\description{
A dataset containing the relative response of the SDSS i-band filter across
the relevent wavelength range in Angstroms.
}
\keyword{datasets}
