% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filt_u_SDSS}
\alias{filt_u_SDSS}
\title{u-band SDSS filter responses at a range of wavelengths}
\format{
A data table with 2 variables and 47 observations:
\describe{
\item{wave}{a numeric representing the wavelength at which the relative response is recorded, given in Angstroms}
\item{response}{a numeric representing the relative response of the filter at that wavelength}
}
}
\source{
\url{https://classic.sdss.org/dr7/instruments/imager/filters/u.dat}
}
\usage{
filt_u_SDSS
}
\description{
A dataset containing the relative response of the SDSS u-band filter across
the relevent wavelength range in Angstroms.
}
\keyword{datasets}
