% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_r50.R
\name{find_r50}
\alias{find_r50}
\title{Find the half mass radius of a simulated galaxy.}
\usage{
find_r50(filename, ptype = NA, r200 = 200, fract = 0.5)
}
\arguments{
\item{filename}{The Gadget file containing the particle information of the galaxy to be analysed.}

\item{ptype}{The particle type/types to be extracted - NA (default) gives all particles in the
simulation, 1 - gas, 2 - dark matter, 3 - disc, 4 - bulge, 5 - stars, 6 - boundary.}

\item{r200}{The virial radius specified in the simulation, kpc.}

\item{fract}{The fraction of particles to be contained within the radius calculated. Default is
0.5, i.e. R50.}
}
\value{
Returned is a data.frame that contains the half mass radius, the number of particles
contained within that radius, the total number of particles in the simulation and the exact
fraction of particles contained within that radius.
}
\description{
The purpose of this function is to find the half mass radius of a simulated galaxy. It works by
growing a sphere and returning the radial coordinate at which half the total number of particles
in the galaxy is contained.
}
\examples{
data      = obs_data_prep(filename = system.file("extdata", 'S0_vignette', package="SimSpin"))
ifucube   = ifu_cube(obs_data = data)

output = find_r50(filename     = system.file("extdata", 'S0_vignette', package="SimSpin"),
                  ptype        = NA,
                  r200         = 20)

}
