% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reff_3d.R
\name{find_reff_3d}
\alias{find_reff_3d}
\title{Find the 3D effective radius of a simulated galaxy.}
\usage{
find_reff_3d(filename, ptype = NA, r200 = 200, fract = 0.5)
}
\arguments{
\item{filename}{The Gadget file containing the particle information of the galaxy to be analysed.}

\item{ptype}{The particle type/types to be extracted - NA (default) gives all particles in the simulation, 1 - gas, 2 - dark matter, 3 - disc, 4 - bulge, 5 - stars, 6 - boundary.}

\item{r200}{The virial radius specified in the simulation, kpc.}

\item{fract}{The fraction of particles to be contained within the radius calculated. Default is 0.5, i.e. Reff.}
}
\value{
Returned is a scaled axis ratio that describes the semimajor and semiminor axes of an ellipse that contains half the total number of particles.
}
\description{
The purpose of this function is to find the observed effective radius of a simulated galaxy projected at a user supplied inclination. It works
by ordering particles by their observed radius and returning the radial coordinate at which half the total number of particles in the galaxy is contained.
}
\examples{
\dontrun{
find_reff(filename = "path/to/some/snapshot_XXX",
          ptype    = 0,
          r200     = 10,
          inc_deg  = 0)
}

}
