% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_grid.R
\name{flux_grid}
\alias{flux_grid}
\title{Generating the fluxes for each element of the IFU data-cube.}
\usage{
flux_grid(obs_data, filter = "g")
}
\arguments{
\item{obs_data}{The list output from the \code{obs_data_prep()} function.}

\item{filter}{If Age/Metallicity is supplied, the filter within which the SED is generated.
Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.}
}
\value{
Returns a list containing:
\item{\code{$flux_grid}}{A 3D array of flux in Jansky.}
\item{\code{$image_grid}}{A list containing the indices of particles in each spaxel position.}
}
\description{
The purpose of this function is to construct the mock flux values within each cell of the IFU
cube. It accepts output parameters from \code{obs_data_prep()} and returns a 3D array containing
the flux at each cell position due to contributions from the particles. If particle ages and
metallicities are supplied, an SED is generated in each cell using ProSpect. Else, the
luminosity in each cell is converted to flux.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
data        = obs_data_prep(simdata = galaxy_data)
flux_img    = flux_grid(obs_data  = data)

}
