% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifu_img.R
\name{ifu_img}
\alias{ifu_img}
\title{Creating a set of mock IFU observational images.}
\usage{
ifu_img(obs_data, threshold)
}
\arguments{
\item{obs_data}{The list output from the \code{obs_data_prep()} function.}

\item{threshold}{The minimum number of counts in the image.}
}
\value{
Returns a list containing a mock IFU data images including a counts image (\code{$counts_img}), a velocity image (\code{$velocity_img})
and a dispersion image (\code{$dispersion_img}) as required for calculating the observed spin parameter. Also contained is an image that
describes the shape of the apperture (\code{$appregion}) such that any further convolutions that are applied to mimic beam smearing or seeing
can be trimmed to the appropriate apperture shape.
}
\description{
The purpose of this function is to construct an IFU images. It accepts output parameters from \code{obs_data_prep()} and returns
a counts image, a velocity image and a dispersion image along with an apperture region image that describes the shape of the apperture.
}
\examples{
\dontrun{
data = obs_data_prep()

ifu_img(obs_data  = data,
        threshold = 0)

ifu_img(obs_data  = data,
        threshold = 20)
}

}
