% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_grid.R
\name{img_grid}
\alias{img_grid}
\title{Creating a mock flux image.}
\usage{
img_grid(obs_data, z = 0.05, filter = "g")
}
\arguments{
\item{obs_data}{The list output from the \code{obs_data_prep()} function.}

\item{z}{The projected redshift of the observation, as used in the \code{obs_data_prep()}
function.}

\item{filter}{If Age/Metallicity is supplied, the filter within which the SED is generated.
Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.}
}
\value{
Returns a list containing:
\item{\code{$flux}}{A 2D array of flux in Jansky.}
\item{\code{$image_grid}}{A list containing the indices of particles in each spaxel position.}
}
\description{
The purpose of this function is to construct a mock flux image. It accepts output parameters from
\code{obs_data_prep()} and returns a 2D array containing the flux at each spaxel position due to
contributions from the particles. If particle ages and metallicities are supplied, an SED is
generated in each spaxel using ProSpect. Else, the luminosity in each cell is converted to flux.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
data        = obs_data_prep(simdata = galaxy_data)
flux_img    = img_grid(obs_data  = data)

}
