% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_correct.R
\name{inc_correct}
\alias{inc_correct}
\title{Inclination correction for \eqn{\lambda}_R.}
\usage{
inc_correct(inc_deg, galaxy_class, axis_ratio, obs_lambdar, v_aniso = c(0,
  0.2, 0.4, 0.6))
}
\arguments{
\item{inc_deg}{The true inclination at which the galaxy was observed (as specified in \code{obs_data_prep()}).}

\item{galaxy_class}{The Hubble type of the galaxy with options \code{c("early","late")}.}

\item{axis_ratio}{A data frame containing the semi-major and semi-minor axes lengths for the observed galaxy, as given by ifu_cube().}

\item{obs_lambdar}{The observed \eqn{\lambda}_R value as produced by the \code{obs_lambda()} function.}

\item{v_aniso}{\emph{Optional} parameter that specifies the velocity anisotropy at which the inclination correction is made.
If missing, the correction will be applied for \code{v_aniso = c(0, 0.2, 0.4, 0.6)}.}
}
\value{
Returns a list containing the observed inclination (\code{$obs_inc}) and a matrix containing the \eqn{\lambda}_R values (\code{$obs_lambda})
for a true inclination correction (\code{"true_inc_correct"}) and the observed inclination correction (\code{"obs_inc_correct"}) for each velocity
anisotropy.
}
\description{
The purpose of this function is to take the \eqn{\lambda}_R value observed via the \code{obs_lambda()} function and correct
it for inclination, as done for true IFU observations. You supply both the true and observed inclinations, as calculated
by the \code{obs_inclination()} function. The true anisotropy can also be supplied, as calculated from the simulation data
in \code{sim_analysis()}. Else, the inclination correction will be calculated for 4 anisotropy values between 0 and 0.6.
}
\examples{
\dontrun{
images      = ifu_cube()
lambda_data = obs_lambda()

inc_correct(inc_deg      = 0,
            galaxy_class = "late",
            axis_ratio   = images$axis_ratio,
            obs_lambdar  = lambda_data$obs_lambdar)

inc_correct(inc_deg      = 0,
            galaxy_class = "early",
            axis_ratio   = images$axis_ratio,
            obs_lambdar  = lambda_data$obs_lambdar,
            v_aniso      = 0.1)

inc_correct(inc_deg      = 0,
            galaxy_class = "late",
            axis_ratio   = images$axis_ratio,
            obs_lambdar  = lambda_data$obs_lambdar,
            v_aniso     = c(0, 0.1, 0.2, 0.3))
}

}
