% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{kpc_per_arcsec,Distance-method}
\alias{kpc_per_arcsec,Distance-method}
\title{Query the kpc_per_arcsec of a Distance object}
\usage{
\S4method{kpc_per_arcsec}{Distance}(x)
}
\arguments{
\item{x}{Distance object}
}
\value{
The distance to the projected galaxy in units of kilo-parsec per
 arcsecond.
}
\description{
Query the kpc_per_arcsec of a Distance object
}
