% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{Mpc,Distance-method}
\alias{Mpc,Distance-method}
\title{Query the Mpc of a Distance object}
\usage{
\S4method{Mpc}{Distance}(x)
}
\arguments{
\item{x}{Distance object}
}
\value{
The distance to the projected galaxy in units of Mpc.
}
\description{
Query the Mpc of a Distance object
}
