% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{A class to describe the basic properties of the object being observed}
\usage{
objective(
  dist_z = 0.05,
  dist_Mpc,
  dist_kpc_per_arcsec,
  inc_deg = 70,
  twist_deg = 0,
  pointing_kpc = c(0, 0),
  pointing_deg,
  blur = F,
  fwhm = 2,
  psf = "Gaussian"
)
}
\arguments{
\item{dist_z}{The projected distance at which the galaxy is placed in terms of
redshift distance. This parameter is one of three optional ways of describing
the distance to the observed galaxy. Only ONE of these three must be
specified.}

\item{dist_Mpc}{The projected distance at which the galaxy is placed in terms
of physical distance in units of Mega-parsec. This parameter is the second of
three optional ways of describing the distance to the observed galaxy. Only
ONE of these three must be specified.}

\item{dist_kpc_per_arcsec}{The projected distance at which the galaxy is
placed in terms of angular distance in units of kilo-parsec per arcsecond.
This parameter is the third of three optional ways of describing the distance
to the observed galaxy. Only ONE of these three must be specified.}

\item{inc_deg}{Numeric describing the projected inclination of the observed
galaxy relative to the z-axis - 0 deg places the galaxy face-on, 90 deg is
edge-on aligned with the horizontal axis. Default is 70.}

\item{twist_deg}{Numeric describing the viewer's orientation relative to the
x-axis - 0 deg places the galaxy face-on, 90 deg places the galaxy edge-on
aligned with the vertical axis. Default is 0.}

\item{pointing_kpc}{Numeric array `c(x,y)`. Two elements specifying the
position at which the observation is centred given as a shift relative to the
centre at (0,0) in units of kilo-parsecs. This parameter is ONE of two
optional ways of describing this offset. Default is c(0,0) offset.}

\item{pointing_deg}{Numeric array `c(x,y)`. Two elements specifying the
position at which the observation is centred given as a shift relative to the
centre at (0,0) in units of degrees. This parameter is ONE of two
optional ways of describing this offset. Default is c(0,0) offset.}

\item{blur}{Boolean describing whether seeing conditions should be applied.
Default is FALSE.}

\item{fwhm}{If \code{blur = TRUE}, a numeric that describes the full-width
half-maximum of the point spread function (PSF) used to blur the image,
given in arcsec.}

\item{psf}{A string to describe the shape of the PSF. Options include
"Gaussian" or "Moffat". Input is NOT case sensitive.}
}
\value{
Returns an object of class "objective" that describes the
conditions in which the observation is made. Required to run
\code{build_datacube()}.
}
\description{
The purpose of this function is to generate a class that describes the
conditions of the observation for \code{build_datacube()}.
}
\examples{
conditions = objective()

}
