% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_vsigma.R
\name{obs_vsigma}
\alias{obs_vsigma}
\title{Calculating the observable V/\eqn{\sigma}}
\usage{
obs_vsigma(ifu_datacube, reff_axisratio, sbinsize)
}
\arguments{
\item{ifu_datacube}{The list output from the function \code{ifu_cube()} containing the mock IFU cube and the apperture region image (\code{$appregion}).}

\item{reff_axisratio}{The semi-major and semi-minor axes output from the \code{find_reff()} function.}

\item{sbinsize}{The size of each spatial bin in kpc, output from the function \code{obs_data_prep()}.}
}
\value{
Returns a list that contains the observed V/\eqn{\sigma} value (\code{$obs_vsigma}), and three matricies reflecting the images produced
in IFU surveys - a luminosity counts image (\code{$counts_img}), a velocity image (\code{$velocity_img}) and a velocity dispersion image
(\code{$dispersion_img}) - and the coordinates of the effective radius ellipse within which V/\eqn{\sigma}  is measured.
}
\description{
The purpose of this function is to calculate the V/\eqn{\sigma}. You can either supply the cube created
by the \code{ifu_cube()} function directly, or the blurred cube created by \code{blur_cube()}.
}
\examples{
\dontrun{
 data      = obs_data_prep()
 ifucube   = ifu_cube()
 blurcube  = blur_cube()
 reff_data = find_reff()

 obs_vsigma(ifu_datacube   = ifucube,
            reff_axisratio = reff_data,
            sbinsize       = data$sbinsize)

}

}
