% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{observation}
\alias{observation}
\title{A class to describe the basic properties of the observation}
\usage{
observation(telescope, observing_strategy)
}
\arguments{
\item{telescope}{A \code{telescope} object.}

\item{observing_strategy}{An \code{observing_strategy} object.}
}
\value{
Returns an object of class "observation" that summarises the
properties of the observation. Run within the \code{build_datacube()}
function.
}
\description{
The purpose of this function is to generate an object that describes how the
galaxy has been observed.
}
\examples{
sami = telescope(type="SAMI")
conditions = observing_strategy()
observation_summary = observation(telescope=sami, observing_strategy=conditions)

}
