% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observing_strategy.R
\name{observing_strategy}
\alias{observing_strategy}
\title{A class to describe the basic properties of the observing conditions}
\usage{
observing_strategy(z = 0.1, inc_deg = 70, blur = F, fwhm = 2, psf = "Gaussian")
}
\arguments{
\item{z}{Numeric describing the redshift to which the observed galaxy is
projected. Default is 0.1.}

\item{inc_deg}{Numeric describing the projected inclination of the observed
galaxy - 0 deg places the galaxy face-on, 90 deg is edge-on. Default is 70.}

\item{blur}{Boolean describing whether seeing conditions should be applied.
Default is FALSE.}

\item{fwhm}{If \code{blur = TRUE}, a numeric that describes the full-width
half-maximum of the point spread function (PSF) used to blur the image,
given in arcsec.}

\item{psf}{A string to describe the shape of the PSF. Options include
"Gaussian" or "Moffat". Input is NOT case sensitive.}
}
\value{
Returns an object of class "observing_strategy" that describes the
conditions in which the observation is made. Required to run
\code{build_datacube()}.
}
\description{
The purpose of this function is to generate a class that describes the
conditions of the observation for \code{build_datacube()}.
}
\examples{
conditions = observing_strategy()

}
