% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ifu.R
\name{plot_ifu}
\alias{plot_ifu}
\title{Plotting the flux, velocity and dipsersion image output from the IFU kinematic cube.}
\usage{
plot_ifu(lambda, appregion)
}
\arguments{
\item{lambda}{The list output from the function \code{obs_lambda()} containing the flux, velocity and dispersion images.}

\item{appregion}{The aperture region output from the \code{obs_data_prep()} function.}
}
\value{
Returns three plots - a flux image, a velcoity image and a dispersion image.
}
\description{
The purpose of this function is to plot the useful images expracted from the IFU cube produced. Input the apperture region and images produced by the
\code{obs_lambda()} function and the output will be the flux, velocity and dispersion maps.
}
\examples{
\dontrun{
data = obs_data_prep()
lambda_obs = obs_lambda()

plot_ifu(lambda_obs, data$appregion)
}

}
