% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_images.R
\name{plot_voronoi}
\alias{plot_voronoi}
\title{Plotting pretty Voronoi bin images}
\usage{
plot_voronoi(
  voronoi_bins,
  fig = c(0, 1, 0, 1),
  new = F,
  main = "",
  na.color = "white",
  zlim = NA,
  ...
)
}
\arguments{
\item{voronoi_bins}{Numeric array containing the metallicity image.}

\item{fig}{Numeric array of length 4 describing the boundary of the image}

\item{new}{Boolean. Should the image be added to the existing plot? Default
is FALSE.}

\item{main}{Image title, default "".}

\item{na.color}{String. Colour given to NA values in the image.}

\item{zlim}{Numeric array of length 2. Describing the numeric range of
colours in the image. Default is NA, in which the range will be described by
the minimum and maximum values in the image.}

\item{...}{Further variables passed to magimage. See
\code{\link[magicaxis]{magimage}} for further details.}
}
\value{
Returns an image to the plotting window of the input
\code{build_datacube} image.
}
\description{
A function to produce a plot of the Voronoi bins produced by
\code{build_datacube()} with associated colour bar and labels.
}
\examples{
ss_pd_eagle = system.file("extdata", "SimSpin_example_EAGLE.hdf5",
package = "SimSpin")
ss_eagle = make_simspin_file(ss_pd_eagle, write_to_file = FALSE)
cube = build_datacube(simspin_file = ss_eagle,
                      telescope = telescope(type="SAMI"),
                      observing_strategy = observing_strategy(),
                      method = "velocity", voronoi_bin = T)
plot_voronoi(cube$raw_images$voronoi_bins)
}
