% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pointing.R
\docType{class}
\name{Pointing-class}
\alias{Pointing-class}
\alias{xy_deg}
\alias{xy_kpc}
\title{An S4 class to represent the telescope pointing relative to the galaxy centre}
\usage{
xy_deg(x)

xy_kpc(x)
}
\arguments{
\item{x}{Pointing object}
}
\description{
An S4 class to represent the telescope pointing relative to the galaxy centre
}
\section{Functions}{
\itemize{
\item \code{xy_deg}: Query the xy_deg of a Pointing object

\item \code{xy_kpc}: Query the xy_kpc of a Pointing object
}}

\section{Slots}{

\describe{
\item{\code{xy_deg}}{Numeric of length 2. Describes the pointing relative to the
galaxy centre in units of degrees.}

\item{\code{xy_kpc}}{Numeric of length 2. Describes the pointing relative to the
galaxy centre in units of kilo-parsecs.}
}}

