% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pointing.R
\name{Pointing}
\alias{Pointing}
\title{A class to describe the center pointing of the telescope relative to the
centre of the observed object.
The purpose of this function is to generate an object that contains the
pointing of the centre of the observation relative to the centre of the
observed galaxy in a variety of units. Supply a single pair of inputs (i.e.
pointing in units of physical distance, "x_kpc" and "y_kpc" OR angular
distance, "x_deg" and "y_deg") and all others will be contained
by the object.}
\usage{
Pointing(xy_deg, xy_kpc, distance)
}
\arguments{
\item{xy_deg}{Offset from centre (0,0) in angular units of degrees along the
x-axis and y-axis respectively.}

\item{xy_kpc}{Offset from centre (0,0) in  physical units of kilo-parsec along
the x-axis and y-axis respectively.}

\item{distance}{The Distance object used to describe the projected distance
of the observed galaxy.}
}
\value{
Returns an object of class "Pointing" that summarises the
pointing position of the telescope relative to the centre of the object
being observed in all associated units.
}
\description{
A class to describe the center pointing of the telescope relative to the
centre of the observed object.
The purpose of this function is to generate an object that contains the
pointing of the centre of the observation relative to the centre of the
observed galaxy in a variety of units. Supply a single pair of inputs (i.e.
pointing in units of physical distance, "x_kpc" and "y_kpc" OR angular
distance, "x_deg" and "y_deg") and all others will be contained
by the object.
}
\examples{
dist = Distance(z=0.3)
Pointing(xy_kpc = c(-1,0), distance = dist)

}
