% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_galaxy}
\alias{r_galaxy}
\title{Computing radius.}
\usage{
r_galaxy(part_data)
}
\arguments{
\item{part_data}{A data.frame describing the particles ID, positions,
velocities and masses.}
}
\value{
The radius of the particle positions.
}
\description{
The purpose of this function is to compute the radial coordinates.
}
\examples{
  galaxy_data = data.frame("ID"=1:100, "x"=stats::rnorm(100),
                           "y"=stats::rnorm(100), "z"=stats::rnorm(100),
                           "vx"=stats::rnorm(100), "vy"=stats::rnorm(100),
                           "vz"=stats::rnorm(100), "Mass"=rep(1,100))
  r = r_galaxy(galaxy_data)
}
