% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Prepare simulation data for SimSpin}
\usage{
sim_data(filename, ptype = NA, SSP = NA, m2l_disc = 1,
  m2l_bulge = 1, m2l_star = 1)
}
\description{
Prepare simulation data for SimSpin
}
