% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_FITS.R
\name{sim_FITS}
\alias{sim_FITS}
\title{Generate a FITS file for the data cube produced in SimSpin}
\usage{
sim_FITS(out_data, out_file, obs_name = "SimSpin datacube")
}
\arguments{
\item{out_data}{The list output from \code{\link{build_datacube}}, \code{\link{find_lambda}} or
\code{\link{find_vsigma}}.}

\item{out_file}{A string describing the path and file name of the FITS file to be written.}

\item{obs_name}{A string that describes the name of the observation. Default is "SimSpin datacube".}
}
\value{
Outputs a standard format FITS file.
}
\description{
The purpose of this function is to write out the data cube produced by SimSpin into the common
astronomy FITS data format.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
lambdar = find_lambda(simdata      = galaxy_data,
                      r200         = 200,
                      z            = 0.1,
                      fov          = 15,
                      ap_shape     = "circular",
                      central_wvl  = 4800,
                      lsf_fwhm     = 2.65,
                      pixel_sscale = 0.5,
                      pixel_vscale = 1.04,
                      inc_deg      = 0,
                      threshold    = 25,
                      measure_type = list(type = "fixed",
                                          axis_ratio = data.frame("a"=3.5, "b"=1.7, "angle"=90),
                                          fac = 1),
                      IFU_plot     = FALSE)
sim_FITS(out_data = lambdar, out_file = "simdata_example.fits")
unlink("simdata_example.fits")
}
