% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_losvd.R
\name{sim_losvd}
\alias{sim_losvd}
\title{Kinematic analysis of the LOSVD for simulation data.}
\usage{
sim_losvd(filename, inc_deg, rmax = 200, ptype = NA, los_upper = c(1, 1),
  los_lower = c(-1, -1), plot_out = TRUE)
}
\arguments{
\item{filename}{The Gadget file containing the particle information of the galaxy to be analysed.}

\item{inc_deg}{The inclination at which to observe the galaxy in degrees.}

\item{rmax}{The maximum radius to consider when constructing the LOSVD.}

\item{ptype}{The particle type/types to be extracted - \code{NA} (default) gives all particles in
the simulation, 1 - gas, 2 - dark matter, 3 - disc, 4 - bulge, 5 - stars, 6 - boundary.}

\item{los_upper}{The upper boundary for the L-BFGS-B fit.}

\item{los_lower}{The lower boundary for the L-BFGS-B fit.}

\item{plot_out}{Logical specifying whether or not you would like the function to output a plot. 
Default is \code{TRUE}.}
}
\value{
A list that contains the particle data of the simulation (\code{$part_data} and 
\code{$head_data} for a particle data.frame and header file respectively), a data frame 
containing the projected data, and the best fit h$_3$ and h$_4$ coefficients.
}
\description{
The purpose of this function is to fit the parameters of the line-of-sight velocity distribution
(LOSVD) and return the gauss-hermite coefficients, h$_3$ and h$_4$.
}
