% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{twist_galaxy}
\alias{twist_galaxy}
\title{Twisting the galaxy relative to the observer.}
\usage{
twist_galaxy(part_data, twist_rad)
}
\arguments{
\item{part_data}{A data.frame describing the particles positions and
velocities.}

\item{twist_rad}{The observed inclination angle in radians.}
}
\value{
Returns the original data frame containing the original particle
 information but in the new orientation.
}
\description{
The purpose of this function is to re-orient the galaxy and rotate the
observer around the z-axis.
}
\examples{
  galaxy_data = data.frame("x"=stats::rnorm(100), "y"=stats::rnorm(100),
                           "z"=stats::rnorm(100), "vx"=stats::rnorm(100),
                           "vy"=stats::rnorm(100), "vz"=stats::rnorm(100))
  observed_data = twist_galaxy(galaxy_data, twist_rad = 1.047)
}
