% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_simspin_FITS.R
\name{write_simspin_FITS}
\alias{write_simspin_FITS}
\title{A function for making a mock spectral cube FITS file output}
\usage{
write_simspin_FITS(output_file, simspin_cube, observation)
}
\arguments{
\item{output_file}{The path to the location where the FITS file should be
written.}

\item{simspin_cube}{The spectral data cube output when \code{build_datacube()}
is run.}

\item{observation}{An object of the observation class, as generated when
\code{build_datacube()} is run.}
}
\value{
Returns an .fits file that contains a the generated spectral cube and
relevant header describing the mock observation.
}
\description{
The purpose of this function is to write the spectral cube generated using
the \code{build_datacube()} function to a FITS file.
}
\examples{
ss_eagle = system.file("extdata", "SimSpin_example_EAGLE.hdf5", package = "SimSpin")
temp_loc = tempdir()
make_simspin_file(ss_eagle, output = paste(temp_loc, "spectra.fst", sep=""))
cube = build_datacube(simspin_file = paste(temp_loc, "spectra.fst", sep=""),
                     telescope = telescope(type="SAMI"),
                     observing_strategy = observing_strategy())
write_simspin_FITS(output_file = paste(temp_loc, "cube.fits", sep=""),
                  simspin_cube = cube$spectral_cube,
                  observation = cube$observation)
unlink(paste(temp_loc, "spectra.fst", sep=""))
unlink(paste(temp_loc, "cube.fits", sep=""))
}
