% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pointing.R
\name{xy_deg,Pointing-method}
\alias{xy_deg,Pointing-method}
\title{Query the xy_deg of a Pointing object}
\usage{
\S4method{xy_deg}{Pointing}(x)
}
\arguments{
\item{x}{Pointing object}
}
\value{
The (x,y) offset (in units of degrees) of the telescope pointing
 relative to the galaxy centre.
}
\description{
Query the xy_deg of a Pointing object
}
