% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{z,Distance-method}
\alias{z,Distance-method}
\title{Distance Query the z of a Distance object}
\usage{
\S4method{z}{Distance}(x)
}
\arguments{
\item{x}{Distance object}
}
\value{
The distance to the projected galaxy in units of redshift.
}
\description{
Distance Query the z of a Distance object
}
