% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur_cube.R
\name{blur_cube}
\alias{blur_cube}
\title{Mimicking seeing conditions for the IFU data cube.}
\usage{
blur_cube(ifu_datacube, sbinsize, psf, fwhm, angular_size)
}
\arguments{
\item{ifu_datacube}{The list output from the function \code{ifu_cube()} containing the mock IFU cube and the apperture region image (\code{$appregion}).}

\item{sbinsize}{The size of each spatial bin in kpc, output from the function \code{obs_data_prep()}.}

\item{psf}{This parameter gives the user choice between a "Gaussian" or "Moffat" PSF. The default is Gaussian.}

\item{fwhm}{This parameter will blur the observation using the specified PSF with full width half maximum as
specified here in arcseconds, in order to mimic the effects caused by beam smearing and seeing.}

\item{angular_size}{This parameter only needs to be specified if a fwhm is given. The conversion factor for a given redshift between
kpc and arcseconds as produced by the \code{obs_data_prep()} function.}
}
\value{
Returns a list that contains the observed \eqn{\lambda}_R value (\code{$obs_lambdar}), and three matricies reflecting the images produced
in IFU surveys - a luminosity counts image (\code{$counts_img}), a velocity image (\code{$velocity_img}) and a velocity dispersion image
(\code{$dispersion_img}) - and the coordinates of the effective radius ellipse within which \eqn{\lambda}_R is measured.
}
\description{
The purpose of this function is to mimic the spatial blurring caused by seeing and beam smearing.
}
\examples{
\dontrun{
 data      = obs_data_prep()
 ifucube   = ifu_cube()

 blur_cube(ifu_datacube   = ifucube,
           sbinsize       = data$sbinsize,
           psf            = "Moffat",
           fwhm           = 0.5,
           angular_size   = data$angular_size)
}
}
