% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur_cube.R
\name{blur_cube}
\alias{blur_cube}
\title{Mimicking seeing conditions for the IFU data cube.}
\usage{
blur_cube(ifu_datacube, sbinsize, psf, fwhm, angular_size)
}
\arguments{
\item{ifu_datacube}{The list output from the function \code{\link{ifu_cube}} containing the mock
IFU cube and the aperture region image (\code{$appregion}).}

\item{sbinsize}{The size of each spatial bin in kpc, output from the function
\code{\link{obs_data_prep}}.}

\item{psf}{The type of PSF with choices of "Gaussian" or "Moffat".}

\item{fwhm}{The full-width half-maximum size of the chosen PSF specified here in arcseconds.}

\item{angular_size}{The conversion factor for a given redshift between kpc and arcseconds as
produced by the \code{\link{obs_data_prep}} function.}
}
\value{
Returns a list that contains:
\item{\code{$blurcube}}{The original IFU cube produced by the function \code{\link{ifu_cube}},
but blurred with the specified PSF.}
\item{\code{$xbin_labels}}{Bin labels for the x-spatial dimension.}
\item{\code{$ybin_labels}}{Bin labels for the y-spatial dimension.}
\item{\code{$vbin_labels}}{Bin labels for the velocity dimension.}
}
\description{
The purpose of this function is to mimic the spatial blurring caused by seeing and beam smearing.
}
\examples{
 galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
 data      = obs_data_prep(galaxy_data)
 ifucube   = ifu_cube(obs_data = data)

 blurcube  = blur_cube(ifu_datacube   = ifucube,
                       sbinsize       = data$sbinsize,
                       psf            = "Moffat",
                       fwhm           = 0.5,
                       angular_size   = data$angular_size)

}
