% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur_cube.R
\name{blur_cube}
\alias{blur_cube}
\title{Mimicking seeing conditions for the IFU data cube.}
\usage{
blur_cube(obs_data, ifu_datacube, psf, fwhm)
}
\arguments{
\item{obs_data}{The list output from the \code{obs_data_prep()} function.}

\item{ifu_datacube}{The list output from the function \code{\link{ifu_cube}} containing the mock
IFU cube and the aperture region image (\code{$ap_region}).}

\item{psf}{The type of PSF with choices of "Gaussian" or "Moffat".}

\item{fwhm}{The full-width half-maximum size of the chosen PSF specified here in arcseconds.}
}
\value{
Returns a list that contains:
\item{\code{$blurcube}}{The original IFU cube produced by the function \code{\link{ifu_cube}},
 but blurred with the specified PSF.}
\item{\code{$ap_region}}{An image that describes the shape of the aperture such that any further
 convolutions that are applied to mimic beam smearing or seeing can be trimmed to the
 appropriate aperture shape.}
}
\description{
The purpose of this function is to mimic the spatial blurring caused by seeing and beam smearing.
}
\examples{
 galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
 data        = obs_data_prep(galaxy_data)
 fluxes      = flux_grid(obs_data = data)
 cube        = ifu_cube(obs_data  = data, flux_data = fluxes)

 blurcube  = blur_cube(ifu_datacube   = ifucube,
                       sbinsize       = data$sbinsize,
                       psf            = "Moffat",
                       fwhm           = 0.5,
                       angular_size   = data$angular_size)

}
