% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur_datacube.R
\name{blur_datacube}
\alias{blur_datacube}
\title{A function for applying seeing conditions to the mock cube}
\usage{
blur_datacube(cube)
}
\arguments{
\item{cube}{The list output from \code{\link{build_datacube}}.}
}
\value{
Returns a new list that contains the spectral cube produced by
\code{\link{build_datacube}}, but with the seeing conditions applied.
}
\description{
The purpose of this function is to apply a convolution kernel to each spatial
plane of the cube produced in \code{\link{build_datacube}}.
}
\examples{
\dontrun{
ss_eagle = system.file("extdata", "SimSpin_example_EAGLE.hdf5", package = "SimSpin")
temp_loc = tempdir()
make_simspin_file(ss_eagle, output = paste(temp_loc, "spectra.fst", sep=""))
cube = build_datacube(simspin_file = paste(temp_loc, "spectra.fst", sep=""),
                     telescope = telescope(type="SAMI"),
                     observing_strategy = observing_strategy())
unlink(paste(temp_loc, "spectra.fst", sep=""))
}

}
