% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datacube.R
\name{build_datacube}
\alias{build_datacube}
\title{A function for making a mock spectral/velocity data cube}
\usage{
build_datacube(
  simspin_file,
  telescope,
  objective,
  verbose = F,
  write_fits = F,
  output_location,
  object_name = "GalaxyID_unknown",
  telescope_name = "SimSpin",
  observer_name = "Anonymous",
  cores = 1,
  mass_flag = F
)
}
\arguments{
\item{simspin_file}{The path to the location of the SimSpin .Rdata file OR
output list from \code{make_simspin_file()}.}

\item{telescope}{An object of the telescope class describing the
specifications of the observing telescope (i.e. field of view, spatial
resolution, wavelength resolution, etc.). See
\code{\link{telescope}} help for more details.}

\item{objective}{An object of the objective class that
describes the properties of the observed simulation (i.e. redshift,
inclination, seeing conditions). See \code{\link{objective}}
help for more details.}

\item{verbose}{Default is \code{FALSE}. If you would like the code to give
updates about its progress, change this parameter to \code{TRUE}.}

\item{write_fits}{Default is \code{FALSE}. If you would like the code to
output a FITS file, change this parameter to \code{TRUE}.}

\item{output_location}{Optional parameter that describes the path to the FITS
file output if \code{write_fits = TRUE}. If \code{write_fits = TRUE} and no
\code{output_location} is specified, the FITS file will be written to the
same directory as the input \code{simspin_file}.}

\item{object_name}{Optional string used in \code{write_simspin_FITS} to
describe the name of the object observed in FITS header.}

\item{telescope_name}{Optional string used in \code{write_simspin_FITS} to
describe the name of the telescope used for observation in FITS header.}

\item{observer_name}{Optional string used in \code{write_simspin_FITS} to
describe the name of the person who ran the observation in FITS header.}

\item{cores}{Float describing the number of cores to run the interpolation
and velocity gridding on. Default is 1.}

\item{mass_flag}{Boolean flag that, when set to TRUE, will compute properties
using a mass weighting rather than a luminosity weighting. Default is FALSE.}
}
\value{
Returns a list containing four elements:
\enumerate{
\item \code{spectral_cube} or \code{velocity_cube} - a 3D array containing
either a spectral cube or a velocity cube, where the output type is
determined by the \code{method} selected in the \code{telescope} function.
\item \code{observation} - a list containing a summary of the details of the
observation (i.e. the output from the function \code{observation()}).
\item \code{raw_images} - a list of 2D arrays, where each 2D array represents
a raw particle image gridded as for the observation details.
\item \code{observed_images} - NULL or a list of 2D arrays (again, where the
output type is determined by the \code{method} selected in the
\code{telescope} function.) containing kinematic images of the collapsed
cube. If \code{blur=T}, these images will be blurred to the specified amount.
}
If \code{write_fits = T}, a .fits file that contains a the generated cube and
relevant header describing the mock observation will also be produced at the
specified \code{output_location}.
}
\description{
The purpose of this function is to generate a mock IFU data cube from
a SimSpin file, as generated using the \code{make_simspin_file()} function.
The data cube produced can be either a spectral cube or a velocity cube.
}
\examples{
ss_gadget = system.file("extdata", "SimSpin_example_Gadget_spectra.Rdata",
                        package = "SimSpin")
cube = build_datacube(simspin_file = ss_gadget,
                     telescope = telescope(type="SAMI"),
                     objective = objective())

}
