% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datacube.R
\name{build_datacube}
\alias{build_datacube}
\title{Constructing kinematic data cubes.}
\usage{
build_datacube(simdata, r200 = 200, z, fov, ap_shape, central_wvl,
  lsf_fwhm, pixel_sscale, pixel_vscale, inc_deg, filter = "g", blur)
}
\arguments{
\item{simdata}{The simulation information data.frame output by \code{\link{sim_data}}.}

\item{r200}{The virial radius specified in the simulation, kpc.}

\item{z}{The galaxy redshift.}

\item{fov}{The field of view of the IFU, diameter in arcseconds.}

\item{ap_shape}{The shape of the field of view, with options "circular", "square" or "hexagonal".}

\item{central_wvl}{The central filter wavelength used for the observation, given in angstroms.}

\item{lsf_fwhm}{The line spread function full-width half-max, given in angstroms.}

\item{pixel_sscale}{The corresponding spatial pixel scale associated with a given telescope
output in arcseconds.}

\item{pixel_vscale}{The corresponding velocity pixel scale associated with a given telescope
filter output in angstroms.}

\item{inc_deg}{The inclination at which to observe the galaxy in degrees.}

\item{filter}{If Age/Metallicity is supplied, the filter within which the SED is generated.
Options include "r" and "g"  for SDSS-r and SDSS-g bands respectively.}

\item{blur}{\emph{Optional} Specified to apply observational seeing effects to the cube. A list
of the form \code{list("psf" = "Moffat", "fwhm" = 0.5)}. \code{"psf"} specifies the shape of the
PSF chosen and may be either \code{"Moffat"} or \code{"Gaussian"}. \code{"fwhm"} is a numeric
specifying the full-width half-maximum of the PSF given in units of arcseconds.}
}
\value{
A list containing:
\item{\code{$datacube}}{A 3D array corresponding to the kinematic data cube.}
}
\description{
The purpose of this basic function is to use the \code{SimSpin} package sub-functions to
construct an IFU kinematic data cube. This function will call each required sub-function
(\code{\link{obs_data_prep}}, \code{\link{flux_grid}}, \code{\link{ifu_cube}},
\code{\link{blur_cube}}) and produce a 3D array corresponding to the kinematic data cube
for that mock observation.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
cube = build_datacube(simdata      = galaxy_data,
                      z            = 0.1,
                      fov          = 15,
                      ap_shape     = "circular",
                      central_wvl  = 4800,
                      lsf_fwhm     = 2.65,
                      pixel_sscale = 0.5,
                      pixel_vscale = 1.04,
                      inc_deg      = 0)

}
