% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cen_galaxy}
\alias{cen_galaxy}
\title{Centering the galaxy.}
\usage{
cen_galaxy(part_data)
}
\arguments{
\item{part_data}{A data.frame describing the particles ID, positions,
velocities and masses.}
}
\value{
A list containing the input data.frame with positions and velocities
 centered, as well as the median coordiantes and velocities of the original
 distribution
}
\description{
The purpose of this function is to centre the galaxy such that the origin of the
system lies at (0,0,0).
}
\examples{
  data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
  galaxy_data = rbind(data$PartType2$Part, data$PartType3$Part)

  output = cen_galaxy(part_data = galaxy_data)
}
