% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lambda.R
\name{find_lambda}
\alias{find_lambda}
\title{Measuring observable galaxy kinematics.}
\usage{
find_lambda(simdata, r200 = 200, z = 0.05, fov = 15,
  ap_shape = "circular", central_wvl = 4800, lsf_fwhm = 2.65,
  pixel_sscale = 0.5, pixel_vscale = 1.04, inc_deg = 70,
  threshold = 25, filter = "g", measure_type = list(type = "fit", fac
  = 1), blur, dispersion_analysis = FALSE, IFU_plot = TRUE)
}
\arguments{
\item{simdata}{The simulation information data.frame output by \code{\link{sim_data}}.}

\item{r200}{The virial radius specified in the simulation, kpc.}

\item{z}{The galaxy redshift.}

\item{fov}{The field of view of the IFU, diameter in arcseconds.}

\item{ap_shape}{The shape of the field of view, with options "circular", "square" or "hexagonal".}

\item{central_wvl}{The central filter wavelength used for the observation, given in angstroms.}

\item{lsf_fwhm}{The line spread function full-width half-max, given in angstroms.}

\item{pixel_sscale}{The corresponding spatial pixel scale associated with a given telescope
output in arcseconds.}

\item{pixel_vscale}{The corresponding velocity pixel scale associated with a given telescope
filter output in angstroms.}

\item{inc_deg}{The inclination at which to observe the galaxy in degrees.}

\item{threshold}{The flux threshold of the observation.}

\item{measure_type}{A list specifying the radius within which \eqn{\lambda_R} is measured. There
are three options for this:
\enumerate{
\item If \code{list("type" = "fit", "fac" = 1)}, \eqn{\lambda_R} is measured in a specified
 multiple of \code{$fac *} \eqn{R_{eff}}, where \eqn{R_{eff}} has been calculated from the
 unblurred galaxy counts image.
\item If \code{list("type" = "specified", "fract" = 0.5, "axis_ratio" = data.frame("a" = 2,
 "b" = 1))}, the effective radius of the galaxy is measured from the unblurred counts image as
 in "fit", but the axis ratio of the grown ellipse is kept at the supplied axis ratio and grown
 until it contains some fraction (\code{$fract}) of the total particles. \eqn{\lambda_R} is
 calculated within this ellipse.
\item Finally, if \code{list("type" = "fixed", "fac" = 1, "axis_ratio" = data.frame("a" = 2,
"b" = 1, "ang" = 90))}, \eqn{\lambda_R} is measured within an ellipse described by the supplied
"axis_ratio" in kpc (or a multiple of that ellipse size given by "fac") at the position angle
"ang" -  no measurement of the effective radius is made, assuming that the supplied values are
determined using another package.
}}

\item{blur}{\emph{Optional} Specify if you wish to apply observational seeing effects to the
cube. A list of the form \code{list("psf" = "Moffat", "fwhm" = 0.5)}. \code{"psf"} specifies
the shape of the PSF chosen and may be either \code{"Moffat"} or \code{"Gaussian"}.
\code{"fwhm"} is a numeric specifying the full-width half-maximum of the PSF given in units of
arcseconds.}

\item{dispersion_analysis}{\emph{Optional} If specified as \code{TRUE}, the code will output the
mean and median values of the LOS velocity dispersion. Default is \code{FALSE}.}

\item{IFU_plot}{\emph{Optional} If specified \code{FALSE}, the function will not output the IFU flux,
LOS velocity and LOS velocity dispersion images. Default is \code{TRUE}, where plots are output
automatically.}
}
\value{
A list containing:
\item{\code{$datacube}}{The 3D array corresponding to the kinematic data cube.}
\item{\code{$xbin_labels}}{Bin labels for the x-spatial dimension.}
\item{\code{$ybin_labels}}{Bin labels for the y-spatial dimension.}
\item{\code{$vbin_labels}}{Bin labels for the velocity dimension.}
\item{\code{$axis_ratio}}{The axis ratio of the observed galaxy in the form of a data frame where
 \code{$a} is the semi-major axis and \code{$b} is the semi-minor axis given in kpc.}
\item{\code{$lambda_r}}{The observed spin parameter \eqn{\lambda_R}}
\item{\code{$counts_img}}{The observed flux image.}
\item{\code{$velocity_img}}{The observed line-of-sight velocity image.}
\item{\code{$dispersion_img}}{The observed line-of-sight velocity dispersion image.}
\item{\code{$angular_size}}{The angular size of the galaxy in kpc/arcecond at the provided
 redshift.}
\item{\code{$sbinsize}}{The size of the spatial bins in kpc.}
\item{\code{$vbinsize}}{The size of the velocity bins in km/s.}
\item{\code{$appregion}}{The aperture region mask used to remove flux outside of the specified
 aperture.}
}
\description{
The purpose of this basic function is to use the \code{SimSpin} package sub-functions to measure
the observable spin parameter, \eqn{\lambda_R}, of a simulated galaxy model. This function will
call each required sub-function (\code{\link{obs_data_prep}}, \code{\link{ifu_cube}},
\code{\link{blur_cube}}, \code{\link{find_reff}}, \code{\link{obs_lambda}} and
\code{\link{plot_ifu}}) and return the \eqn{\lambda_R} within a specified measurement radius.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
lambdar = find_lambda(simdata      = galaxy_data,
                      r200         = 200,
                      z            = 0.1,
                      fov          = 15,
                      ap_shape     = "circular",
                      central_wvl  = 4800,
                      lsf_fwhm     = 2.65,
                      pixel_sscale = 0.5,
                      pixel_vscale = 1.04,
                      inc_deg      = 0,
                      threshold    = 25,
                      measure_type = list(type = "fixed",
                                          axis_ratio = data.frame("a"=3.5, "b"=1.7, "angle"=90),
                                          fac = 1),
                      IFU_plot     = FALSE)

}
