% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reff.R
\name{find_reff}
\alias{find_reff}
\title{Find the effective radius of a simulated galaxy.}
\usage{
find_reff(simdata, r200 = 200, inc_deg, fract = 0.5, axis_ratio,
  angular_size)
}
\arguments{
\item{simdata}{The simulation information data.frame output by \code{\link{sim_data}}.}

\item{r200}{The virial radius specified in the simulation, kpc.}

\item{inc_deg}{The observed inclination of the simulated galaxy in degrees.}

\item{fract}{The fraction of particles to be contained within the radius calculated. Default is
0.5, i.e. Reff.}

\item{axis_ratio}{A data frame containing the semi-major and semi-minor axes lengths for the
observed galaxy, as given by \code{\link{ifu_cube}} or \code{\link{blur_cube}}.}

\item{angular_size}{The kpc/'' scaling factor output by \code{\link{obs_data_prep}} used to give
the axis ratios in both kpc and ''.}
}
\value{
Returned is a scaled axis ratio that describes the semi-major and semi-minor axes of an
ellipse that contains half the total number of particles.
}
\description{
The purpose of this function is to find the observed effective radius of a simulated galaxy
projected at a user supplied inclination. It works by ordering particles by their observed
radius and returning the radial coordinate at which half the total number of particles in the
galaxy is contained.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
data        = obs_data_prep(simdata = galaxy_data)
ifucube     = ifu_cube(obs_data = data)

output = find_reff(simdata      = galaxy_data,
                   r200         = 10,
                   inc_deg      = 0,
                   axis_ratio   = ifucube$axis_ratio,
                   angular_size = data$angular_size)

}
