% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_vsigma.R
\name{find_vsigma}
\alias{find_vsigma}
\title{Measuring observable galaxy kinematics.}
\usage{
find_vsigma(filename, ptype = NA, r200 = 200, z, fov, ap_shape, central_wvl,
  lsf_fwhm, pixel_sscale, pixel_vscale, inc_deg, m2l_disc, m2l_bulge, threshold,
  measure_type = list(type = "fit", fac = 1), blur)
}
\arguments{
\item{filename}{The Gadget file containing the particle information of the galaxy to be analysed.}

\item{ptype}{The particle type/types to be extracted - NA (default) gives all particles in the
simulation, 1 - gas, 2 - dark matter, 3 - disc, 4 - bulge, 5 - stars, 6 - boundary.}

\item{r200}{The virial radius specified in the simulation, kpc.}

\item{z}{The galaxy redshift.}

\item{fov}{The field of view of the IFU, diameter in arcseconds.}

\item{ap_shape}{The shape of the field of view, with options "circular", "square" or "hexagonal".}

\item{central_wvl}{The central filter wavelength used for the observation, given in angstroms.}

\item{lsf_fwhm}{The line spread function full-width half-max, given in angstroms.}

\item{pixel_sscale}{The corresponding spatial pixel scale associated with a given telescope
output in arcseconds.}

\item{pixel_vscale}{The corresponding velocity pixel scale associated with a given telescope
filter output in angstroms.}

\item{inc_deg}{The inclination at which to observe the galaxy in degrees.}

\item{m2l_disc}{The mass-to-light ratio of the disc component in solar units.}

\item{m2l_bulge}{The mass-to-light ratio of the bulge component in solar units.}

\item{threshold}{The flux threshold of the observation.}

\item{measure_type}{A list specifying the radius within which \eqn{\lambda_R} is measured. If
\code{list(type = "fit", fac = 1)}, \eqn{\lambda_R} is measured in a specified multiple of
\eqn{R_{eff}}, where \eqn{R_{eff}} has been calculated from the unblurred galaxy  counts image.
If \code{list(type = "specified", fac = 1, axis_ratio = data.frame("a" = 2, "b" = 1))},
\eqn{\lambda_R} is calculated within some factor of the effective radius, \code{fac *}
\eqn{R_{eff}} with specified axis ratio externally as a data frame in units of kpc.}

\item{blur}{\emph{Optional} Specify if you wish to apply observational seeing effects to the
cube. A list of the form \code{list("psf" = "Moffat", "fwhm" = 0.5)}. \code{"psf"} specifies the
shape of the PSF chosen and may be either \code{"Moffat"} or \code{"Gaussian"}. \code{"fwhm"} is
a numeric specifying the full-width half-maximum of the PSF given in units of arcseconds.}
}
\value{
A list containing:
\item{\code{$datacube}}{A 3D array corresponding to the kinematic data cube.}
\item{\code{$xbin_labels}}{Bin labels for the x-spatial dimension.}
\item{\code{$ybin_labels}}{Bin labels for the y-spatial dimension.}
\item{\code{$vbin_labels}}{Bin labels for the velocity dimension.}
\item{\code{$axis_ratio}}{The axis ratio of the observed galaxy in the form of a data frame
 where \code{$a} is the semi-major axis and \code{$b} is the semi-minor axis given in kpc.}
\item{\code{$vsigma}}{The observed V/\eqn{\sigma}.}
\item{\code{$counts_img}}{The observed flux image.}
\item{\code{$velocity_img}}{The observed line-of-sight velocity image.}
\item{\code{$dispersion_img}}{The observed line-of-sight velocity dispersion image.}
\item{\code{$reff_ellipse}}{The observed measurement radius.}
And optionally, specified by the \code{dispersion_analysis} parameter, the mean and median
values of the line-of-sight velocity dispersion (\code{$dispersion_analysis}). The observational
images will also be plotted.
}
\description{
The purpose of this basic function is to use the \code{SimSpin} package sub-functions to measure
the observable ratio V/\eqn{\sigma} of a simulated galaxy model. This function will call
each required sub-function (\code{\link{obs_data_prep}}, \code{\link{ifu_cube}},
\code{\link{blur_cube}}, \code{\link{find_reff}}, \code{\link{obs_vsigma}} and
\code{\link{plot_ifu}}) and return the V/\eqn{\sigma} ratio within a user specified measurement
radius.
}
\examples{
vsigma = find_vsigma(filename     = system.file("extdata", 'S0_vignette', package="SimSpin"),
                     r200         = 200,
                     z            = 0.1,
                     fov          = 15,
                     ap_shape     = "circular",
                     central_wvl  = 4800,
                     lsf_fwhm     = 2.65,
                     pixel_sscale = 0.5,
                     pixel_vscale = 1.04,
                     inc_deg      = 0,
                     m2l_disc     = 2,
                     m2l_bulge    = 1,
                     threshold    = 25,
                     measure_type = list(type = "fit", fac = 1))


vsigma = find_vsigma(filename        = system.file("extdata", 'S0_vignette', package="SimSpin"),
                     r200            = 200,
                     z               = 0.1,
                     fov             = 15,
                     ap_shape        = "circular",
                     central_wvl     = 4800,
                     lsf_fwhm        = 2.65,
                     pixel_sscale    = 0.5,
                     pixel_vscale    = 1.04,
                     inc_deg         = 0,
                     m2l_disc        = 2,
                     m2l_bulge       = 1,
                     threshold       = 25,
                     measure_type    = list(type = "specified",
                                            axis_ratio = data.frame("a"=3.5, "b"=1.7),
                                            fac = 1),
                     blur            = list("psf" = "Moffat", "fwhm" = 2))

}
