% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifu_cube.R
\name{ifu_cube}
\alias{ifu_cube}
\title{Creating a set of mock IFU observational images.}
\usage{
ifu_cube(obs_data, threshold)
}
\arguments{
\item{obs_data}{The list output from the \code{obs_data_prep()} function.}

\item{threshold}{The flux threshold of the observation.}
}
\value{
Returns a list containing a mock IFU data cube as required for calculating the observed spin parameter. Also contained is an image that
describes the shape of the apperture (\code{$appregion}) such that any further convolutions that are applied to mimic beam smearing or seeing
can be trimmed to the appropriate apperture shape. The bin labels for each dimension are supplied as \code{$xbin_labels}, \code{$ybin_labels}, \code{$vbin_labels}.
The axis ratio of the galaxy is calculated in this fuction and output in the list as \code{$axis_ratio$a} and \code{$axis_ratio$b}.
}
\description{
The purpose of this function is to construct an IFU data cube. It accepts output parameters from \code{obs_data_prep()} and returns
a 3D array containing the spatial and velocity corrdinates of each particle in the galaxy simulation in order to mimic the arrangement
of an IFU data cube.
}
\examples{
\dontrun{
data = obs_data_prep()

ifu_cube(obs_data = data,
        threshold = 0)

ifu_cube(obs_data = data,
        threshold = 20)
}

}
