% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_calc.R
\name{kin_calc}
\alias{kin_calc}
\title{Calculating the observed kinematics, \eqn{\lambda_R} and V/\eqn{\sigma}}
\usage{
obs_kinematics(ifu_datacube, reff_axisratio, sbinsize,
  radius_type = "Both", addSky = FALSE, threshold = 25,
  mag_zero = 8.9, pixel_sscale = 0.5)
kin_calc(obs_data, obs_images, axisratio)
}
\arguments{
\item{ifu_datacube}{The list output from the function \code{\link{ifu_cube}} containing the mock
IFU cube and the apperture region image (\code{$appregion}).}

\item{reff_axisratio}{The semi-major and semi-minor axes output from the \code{\link{find_reff}}
function.}

\item{sbinsize}{The size of each spatial bin in kpc, output from the function
\code{\link{obs_data_prep}}.}

\item{radius_type}{The method of computing radii - "Circular" i.e. \eqn{r^{2} = x^{2} + y{2}} or
"Elliptical" where r is the semi-major axis of the ellipse having an axis ratio \eqn{b/a} on
which the pixel lies, i.e.
\eqn{r^{2} = \frac{x^{2} (1 - \epsilon)^{2} + y^{2}}{(1 - \epsilon)^2}}. Default is "Both" such
that both \eqn{\lambda_R} values are returned.}

\item{addSky}{A boolean to specify whether to add sky noise to the output images. Default is
FALSE. If TRUE, further parameters including \code{mag_threshold} and \code{mag_zero} described
below.}

\item{threshold}{The magnitude limit of the observation.}

\item{mag_zero}{The magnitude zero point with regards to the mangitude system being used (e.g.
AB or Vega).}

\item{pixel_sscale}{The pixel_sscale input for \code{\link{obs_data_prep}} for calculating
the level of sky noise, if requested.}
}
\value{
Returns a list that contains:
\item{\code{$obs_lambdar}}{The observed spin parameter \eqn{\lambda_R} measured with circular
radii. \emph{(When \code{radius_type = "Both"} or \code{"Circular"}.)}}
\item{\code{$obs_elambdar}}{The observed spin parameter \eqn{\lambda_R} measured with elliptical
radii. \emph{(When \code{radius_type = "Both"} or \code{"Elliptical"}.)}}
\item{\code{$obs_vsigma}}{The observed V/\eqn{\sigma} value.}
\item{\code{$counts_img}}{The observed flux image.}
\item{\code{$velocity_img}}{The observed line-of-sight velocity image.}
\item{\code{$dispersion_img}}{The observed line-of-sight velocity dispersion image.}
}
\description{
The purpose of this function is to calculate the spin parameter, \eqn{\lambda_R},
and ratio V/\eqn{\sigma} that would be observed given an IFU data cube. You can either
supply the cube created by the \code{\link{ifu_cube}} function directly, or the blurred
cube created by \code{\link{blur_cube}}.
}
\examples{
 galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
 data        = obs_data_prep(simdata = galaxy_data)
 ifucube     = ifu_cube(obs_data = data, threshold = 20)
 reff_data = find_reff(simdata      = galaxy_data,
                       r200         = 200,
                       inc_deg      = 0,
                       axis_ratio   = ifucube$axis_ratio,
                       angular_size = data$angular_size)

 output = obs_kinematics(ifu_datacube   = ifucube,
                         reff_axisratio = reff_data,
                         sbinsize       = data$sbinsize)

}
