% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_data_prep.R
\name{obs_data_prep}
\alias{obs_data_prep}
\title{Prepare simulation data for observational kinematic analysis.}
\usage{
obs_data_prep(filename, ptype = NA, r200 = 200, z, fov, ap_shape,
  central_wvl, lsf_fwhm, pixel_sscale, pixel_vscale, inc_deg, m2l_disc,
  m2l_bulge)
}
\arguments{
\item{filename}{The Gadget output file containing the particle information of the galaxy to be analysed.}

\item{ptype}{The particle type/types to be extracted - NA (default) gives all particles in the simulation, 1 - gas, 2 - dark matter, 3 - disc,
4 - bulge, 5 - stars, 6 - boundary.}

\item{r200}{The virial radius specified in the simulation, kpc.}

\item{z}{The galaxy redshift.}

\item{fov}{The field of view of the IFU, diameter in arcseconds.}

\item{ap_shape}{The shape of the field of view, with options "circular", "square" or "hexagonal".}

\item{central_wvl}{The central filter wavelength used for the observation, given in angstroms.}

\item{lsf_fwhm}{The line spread function full-width half-max, given in angstroms.}

\item{pixel_sscale}{The corresponding spatial pixel scale associated with a given telescope output in arcseconds.}

\item{pixel_vscale}{The corresponding velocity pixel scale associated with a given telescope filter output in angstroms.}

\item{inc_deg}{The inclination at which to observe the galaxy in degrees.}

\item{m2l_disc}{The mass-to-light ratio of the disc component in solar units.}

\item{m2l_bulge}{The mass-to-light ratio of the bulge component in solar units.}
}
\value{
Returned is a list that contains a data frame of the observed particle information (\code{$galaxy_obs} (which contains the galaxy particle
info from the GADGET file, with added coordinates \code{$z_obs}, \code{$r_obs}, and \code{$vy_obs} and particle fluxes, \code{$flux}, in units of
1e-16 erg s-1 cm-2) and numerical factors including the number of spatial bins (\code{$sbin}), the size of those spatial bins in kpc and arcseconds
(\code{$sbinsize} and \code{pixsize}), the number of velocity bins (\code{$vbin}), the size of those velocity bins in km/s (\code{$vbinsize}), the
gaussian standard deviation of the line spread function in km/s (\code{lsf_size}) and the angular size of the galaxy in kpc/arcecond at the provided
redshift (\code{$angular_size}).
}
\description{
The purpose of this function is to calculate the properties necessary for constructing an IFU observation data cube of a simulated galaxy at a user
specified inclination and redshift.
}
\examples{
\dontrun{
obs_data_prep(filename     = "path/to/some/snapshot_XXX",
              r200         = 200,
              z            = 0.1,
              fov          = 15,
              ap_shape     = "circular",
              central_wvl  = 4800,
              lsf_fwhm     = 2.65,
              pixel_sscale = 0.5,
              pixel_vscale = 1.04,
              inc_deg      = 0,
              m2l_disc     = 2,
              m2l_bulge    = 1)

obs_data_prep(filename     = "path/to/some/snapshot_XXX",
              ptype        = c(3,4),
              r200         = 200,
              z            = 0.1,
              fov          = 15,
              ap_shape     = "hexagonal",
              central_wvl  = 4800,
              lsf_fwhm     = 2.65,
              pixel_sscale = 0.5,
              pixel_vscale = 1.04,
              inc_deg      = 0,
              m2l_disc     = 1,
              m2l_bulge    = 1)
}

}
