% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obs_galaxy}
\alias{obs_galaxy}
\title{Constructing galaxy observation data from the Gadget output file data.}
\usage{
obs_galaxy(part_data, inc_rad)
}
\arguments{
\item{part_data}{A data.frame describing the particles ID, positions, and
velocities.}

\item{inc_rad}{The observed inclination angle in radians.}
}
\value{
Returns a data frame containing the original particle information plus the observed
 z-position (\code{$z_obs}), observed radial position (\code{$r_obs}) and the observed line of
 sight velocity (\code{$vy_obs}) at the given inclination.
}
\description{
The purpose of this function is to produce the observable features of
simulation data when taking mock IFU observations.
}
\examples{
  galaxy_data = data.frame("ID"=1:100, "x"=stats::rnorm(100),
                           "y"=stats::rnorm(100), "z"=stats::rnorm(100),
                           "vx"=stats::rnorm(100), "vy"=stats::rnorm(100),
                           "vz"=stats::rnorm(100), "Mass" = 1:100,
                           "sed_id" = 1:100, "Initial_Mass" = 1:100)
  observed_data = obs_galaxy(galaxy_data, inc_rad = 1.047)
}
