% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_lambda.R
\name{obs_lambda}
\alias{obs_lambda}
\title{Calculating the observable spin parameter, \eqn{\lambda}_R.}
\usage{
obs_lambda(ifu_datacube, reff_axisratio, sbinsize,
  dispersion_analysis = FALSE)
}
\arguments{
\item{ifu_datacube}{The list output from the function \code{ifu_cube()} containing the mock IFU cube and the apperture region image (\code{$appregion}).}

\item{reff_axisratio}{The semi-major and semi-minor axes output from the \code{find_reff()} function.}

\item{sbinsize}{The size of each spatial bin in kpc, output from the function \code{obs_data_prep()}.}

\item{dispersion_analysis}{(Optional) If specified as \code{TRUE}, the code will output the mean and median values of the LOS velocity dispersion. Default is \code{FALSE}.}
}
\value{
Returns a list that contains the observed \eqn{\lambda}_R value (\code{$obs_lambdar}), and three matricies reflecting the images produced
in IFU surveys - a luminosity counts image (\code{$counts_img}), a velocity image (\code{$velocity_img}) and a velocity dispersion image
(\code{$dispersion_img}) - and the coordinates of the effective radius ellipse within which \eqn{\lambda}_R is measured.
}
\description{
The purpose of this function is to calculate the spin parameter that would be observed given an IFU data cube. You can either supply the cube created
by the \code{ifu_cube()} function directly, or the blurred cube created by \code{blur_cube()}.
}
\examples{
\dontrun{
 data      = obs_data_prep()
 ifucube   = ifu_cube()
 blurcube  = blur_cube()
 reff_data = find_reff()

 obs_lambda(ifu_datacube   = ifucube,
            reff_axisratio = reff_data,
            sbinsize       = data$sbinsize)

 obs_lambda(ifu_datacube   = blurcube,
            reff_axisratio = reff_data,
            sbinsize       = data$sbinsize,
            dispersion_analysis  = TRUE)

}

}
