% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{observation}
\alias{observation}
\title{A class to describe the basic properties of the observation}
\usage{
observation(telescope, objective)
}
\arguments{
\item{telescope}{A \code{telescope} object. See
\code{\link{telescope}} help for more details.}

\item{objective}{An \code{objective} object. See
\code{\link{objective}} help for more details.}
}
\value{
Returns an object of class "observation" that summarises the
properties of the observation. Run within the \code{build_datacube()}
function.
}
\description{
The purpose of this function is to generate an object that describes how the
galaxy has been observed.
}
\examples{
sami = telescope(type="SAMI")
conditions = objective()
observation_summary = observation(telescope=sami, objective=conditions)

}
