% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ifu.R
\name{plot_ifu}
\alias{plot_ifu}
\title{Plotting the flux, velocity and dipsersion image output from the IFU kinematic cube.}
\usage{
plot_ifu(lambda, reff = TRUE, which_plots = NA)
}
\arguments{
\item{lambda}{The list output from the function \code{\link{find_lambda}} containing the flux,
velocity and dispersion images.}

\item{reff}{Boolean specifying whether or not you would like the effective radius ellipse to be
plotted over the image. Default is TRUE.}

\item{which_plots}{String describing which plots you wish to show - any combination of "Flux",
"Velocity" or "Dispersion". Default is NA, which will plot all three.}
}
\value{
Returns three plots - a flux image, a velcoity image and a dispersion image.
}
\description{
The purpose of this function is to plot the useful images expracted from the IFU cube produced.
Input the aperture region and images produced by the \code{\link{find_lambda}} function and the
output will be the flux, velocity and dispersion maps.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
lambdar = find_lambda(simdata = galaxy_data)
plot_ifu(lambdar)

}
