% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ifu.R
\name{plot_ifu}
\alias{plot_ifu}
\title{Plotting the flux, velocity and dipsersion image output from the IFU kinematic cube.}
\usage{
plot_ifu(lambda, reff = TRUE)
}
\arguments{
\item{lambda}{The list output from the function \code{\link{find_lambda}} containing the flux,
velocity and dispersion images.}

\item{reff}{Boolean specifying whether or not you would like the effective radius ellipse to be
plotted over the image. Default is TRUE.}
}
\value{
Returns three plots - a flux image, a velcoity image and a dispersion image.
}
\description{
The purpose of this function is to plot the useful images expracted from the IFU cube produced.
Input the aperture region and images produced by the \code{\link{find_lambda}} function and the
output will be the flux, velocity and dispersion maps.
}
\examples{
galaxy_data = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
lambdar = find_lambda(simdata      = galaxy_data,
                      r200         = 200,
                      z            = 0.05,
                      fov          = 15,
                      ap_shape     = "circular",
                      central_wvl  = 4800,
                      lsf_fwhm     = 2.65,
                      pixel_sscale = 0.5,
                      pixel_vscale = 1.04,
                      inc_deg      = 70,
                      threshold    = 25,
                      measure_type = list(type = "fixed",
                                          axis_ratio = data.frame("a"=3.5, "b"=1.7, "angle"=90),
                                          fac = 1),
                      IFU_plot     = FALSE)

plot_ifu(lambdar)

}
