% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ifu.R
\name{plot_ifu}
\alias{plot_ifu}
\title{Plotting the flux, velocity and dipsersion image output from the IFU kinematic cube.}
\usage{
plot_ifu(lambda)
}
\arguments{
\item{lambda}{The list output from the function \code{\link{find_lambda}} containing the flux,
velocity and dispersion images.}
}
\value{
Returns three plots - a flux image, a velcoity image and a dispersion image.
}
\description{
The purpose of this function is to plot the useful images expracted from the IFU cube produced.
Input the aperture region and images produced by the \code{\link{find_lambda}} function and the
output will be the flux, velocity and dispersion maps.
}
\examples{
lambdar = find_lambda(filename     = system.file("extdata", 'S0_vignette', package="SimSpin"),
                      z            = 0.1,
                      fov          = 15,
                      ap_shape     = "circular",
                      central_wvl  = 4800,
                      lsf_fwhm     = 2.65,
                      pixel_sscale = 0.5,
                      pixel_vscale = 1.04,
                      inc_deg      = 0,
                      m2l_disc     = 2,
                      m2l_bulge    = 1,
                      threshold    = 25,
                      IFU_plot     = FALSE)

plot_ifu(lambdar)

}
