% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_images.R
\name{plot_metallicity}
\alias{plot_metallicity}
\title{Plotting pretty metallicity images}
\usage{
plot_metallicity(
  metallicity_image,
  fig = c(0, 1, 0, 1),
  new = F,
  units = expression("log10(Z/Z"[solar] * ")"),
  main = "",
  na.color = "white",
  zlim = NA,
  legend = T,
  radii = NA,
  titleshift = -4,
  labN = 5,
  radii_col = "red",
  ...
)
}
\arguments{
\item{metallicity_image}{Numeric array containing the metallicity image.}

\item{fig}{Numeric array of length 4 describing the boundary of the image}

\item{new}{Boolean. Should the image be added to the existing plot? Default
is FALSE.}

\item{units}{String describing the units of the values contained
in the image.}

\item{main}{Image title, default "".}

\item{na.color}{String. Colour given to NA values in the image.}

\item{zlim}{Numeric array of length 2. Describing the numeric range of
colours in the image. Default is NA, in which the range will be described by
the minimum and maximum values in the image.}

\item{legend}{Boolean to determine if the colour bar axis should be printed
at the bottom of the image. Default is T.}

\item{radii}{list - containing a, b, and ang (if wishing to plot half-mass
radii ellipse).}

\item{titleshift}{Numeric. Describes the distance between the colour bar and
the units. Default is -4.}

\item{labN}{Numeric. Describes the minimum number of numeric labels added to
the colour bar. Default is 5.}

\item{radii_col}{String. Describing the colour of the ellipse drawn if
\code{radii} list is specified. Default is "red".}

\item{...}{Further variables passed to magimage. See
\code{\link[magicaxis]{magimage}} for further details.}
}
\value{
Returns an image to the plotting window of the input
\code{build_datacube} image.
}
\description{
A function to produce a plot of the stellar metallicity image produced by
\code{build_datacube()} with associated colour bar and labels.
}
\examples{
ss_pd_gadget = system.file("extdata", "SimSpin_example_Gadget",
package = "SimSpin")
ss_gadget = make_simspin_file(ss_pd_gadget, write_to_file = FALSE)
cube = build_datacube(simspin_file = ss_gadget,
                      telescope = telescope(type="SAMI"),
                      observing_strategy = observing_strategy(),
                      method = "velocity")
plot_metallicity(cube$raw_images$metallicity_image)
}
