% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_analysis.R
\name{sim_analysis}
\alias{sim_analysis}
\title{A function for measuring the particle properties of the input simulation}
\usage{
sim_analysis(simspin_file, type = "stars", half_mass = NA, bin_breaks = NA)
}
\arguments{
\item{simspin_file}{The path to the location of the SimSpin .Rdata file OR
output list from \code{make_simspin_file()}.}

\item{type}{String "stars" (default) or "gas" to specify which set of
of particles are used in the property calculations.}

\item{half_mass}{If simulation file contains all particles cutout from a box
(rather than just particles from a single galaxy), you can the half-mass
value at which the alignment function is run. Numeric length = 1. Default is
NA, in which case half the total mass of the supplied simulation data is used.}

\item{bin_breaks}{Optional parameter that allows you to specify radial bin
break positions. Default will give bins spaced with varying sized bins:
`seq(0, 9, by=1), seq(12, 51, by=3), seq(61, 101, by=10), seq(151, 501, by=50)`}
}
\value{

}
\description{
The purpose of this function is to measure the true particle properties of the
simulation prior to observation. This function produces a list that includes
a summary of the total galaxy, a summary of the half-mass properties, and a
data.frame containing the radial trends of a series of properties, computed
in a series of spherical shells.
}
\examples{
ss_gadget = system.file("extdata", "SimSpin_example_Gadget_spectra.Rdata",
                        package = "SimSpin")
props = sim_analysis(simspin_file = ss_gadget)

}
