% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_analysis.R
\name{sim_analysis}
\alias{sim_analysis}
\title{A function for measuring the particle properties of the input simulation}
\usage{
sim_analysis(simspin_file, type = "stars")
}
\arguments{
\item{simspin_file}{The path to the location of the SimSpin .Rdata file OR
output list from \code{make_simspin_file()}.}

\item{type}{String "stars" (default) or "gas" to specify which set of
of particles are used in the property calculations.}
}
\value{
Returns a list that contains:
\describe{
  \item{Properties}{list()}
  \item{HalfMassProperties}{list()}
  \item{RadialTrends}{data.frame()}}
  where \code{Properties} include:
  \describe{\item{Type}{Component considered within analysis}
            \item{TotalMass}{Total mass (solar)}
            \item{MeanAge}{Mean age (Gyr)}
            \item{MeanMetallicity}{Mean metallicity (fraction of solar)}
            \item{NumberOfParticles}{Total number of particles}}
  where \code{HalfMassProperties} include:
  \describe{\item{Mass}{Half mass (solar)}
            \item{RadiusCircular}{Circularised radius at half-mass (kpc)}
            \item{RadiusElliptical}{Elliptical radius at half-mass given shapes p & q (kpc)}
            \item{Shape_p}{Axis ratio (b/a) of particles within half-mass}
            \item{Shape_q}{Axis ratio (c/a) of particles within half-mass}}
  and where \code{RadialTrends} include:
  \describe{\item{Radius}{Radial coordinate at the centre of the radial bin (kpc)}
            \item{Mass}{Mass (solar) contained within radial bin}
            \item{CumulativeMass}{Mass (solar) of all particles contained within radius of this bin}
            \item{Density}{Mass density of shell (Msol/kpc^3)}
            \item{Age}{Mean age (Gyr) of particles within radial bin}
            \item{Metallicity}{Mean metallicity (fraction of solar) of particles within radial bin}
            \item{CircularVelocity}{Circular velocity (km/s) due to mass contained within radius of this bin}
            \item{RotationalVelocity}{Mean velocity along circular orbits (km/s) in radial bin}
            \item{RotationalDispersion}{Standard deviation of velocities in circular orbits (km/s) in radial bin}
            \item{VelocityAnisotropy}{Beta parameter of particles within radial bin}
            \item{SpinParameter_Bullock}{Bullock et al (2001) spin parameter measured from all particles contained within radius of this bin}
            \item{SpecificAngularMomentum}{Mass weighted angular momentum, j (kpc km/s), of particles within radial bin}
            \item{Shape_p}{Axis ratio (b/a) of all particles within radius of this bin}
            \item{Shape_q}{Axis ratio (c/a) of all particles within radius of this bin},
            \item{NumberOfParticles}{Number of particles contained within radial bin}}
}
\description{
The purpose of this function is to measure the true particle properties of the
simulation prior to observation. This function produces a list that includes
a summary of the total galaxy, a summary of the half-mass properties, and a
data.frame containing the radial trends of a series of properties, computed
in a series of spherical shells.
}
\examples{
ss_gadget = system.file("extdata", "SimSpin_example_Gadget_spectra.Rdata",
                        package = "SimSpin")
props = sim_analysis(simspin_file = ss_gadget)

}
