% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_analysis.R
\name{sim_analysis}
\alias{sim_analysis}
\title{Kinematic analysis of simulation data.}
\usage{
sim_analysis(simdata, bin_type = "r", rmax = 200, rbin = 200,
  DM_profile = NA)
}
\arguments{
\item{bin_type}{The direction in which to bin the simulation model - \code{"r"} (default) bins
radially in 3D spherical shells, \code{"cr"} bins radially in 2D circular rings, \code{"z"} bins
in 1D off the plane of the galaxy.}

\item{rmax}{The maximum radial coordinate considered within the simulated galaxy in kpc.}

\item{rbin}{The number of radial bins considered.}

\item{DM_profile}{If dark matter particles are not included in the analysis, this option allows
you to use the DM profile for the mass distribution such that the circular velocity can be
correctly determined. Options include \code{NA} (default),
\code{list("profile"="NFW", "DM_vm"=186.9, "DM_a"=34.5, "DM_rhof"=0.035)} - where DM_vm is the virial mass, DM_a is the
scale radius, and DM_rhof is the density evaluated at the flattening radius - and
\code{list("profile"="Hernquist", "DM_mass"=184.9, "DM_a"=34.5)} - where DM_mass is the total mass of the dark
matter component and DM_a is the scale radius of the halo.}

\item{filename}{The Gadget file containing the particle information of the galaxy to be analysed.}

\item{ptype}{The particle type/types to be extracted - NA (default) gives all particles in the
simulation, 0 - gas, 1 - dark matter, 2 - disc, 3 - bulge, 4 - stars, 5 - boundary.}
}
\value{
A list that contains the particle data of the simulation (\code{$part_data} and \code{$head_data} for a particle data.frame and header
file respectively), a data frame containing kinematic features of radial bins (\code{$profile}) including (at least):
\item{\code{$r}/\code{$cr}/\code{$z}}{The outer radius of each bin.}
\item{\code{$Mass}}{The contained mass.}
\item{\code{$logp}}{The logarithm of the density in each bin.}
\item{\code{$vr}/\code{$vcr}/\code{$vz}}{The average velocities in each bin.}
\item{\code{$sigma_vr}/\code{$sigma_vcr}/\code{$sigma_vz}}{The average velocity dispersions in
each bin.}
\item{\code{$J}}{The angular momentum magnitude in each shell.}
In the case of \code{bin_type = "r"} additionally the circular velocity
(\code{$vc}), the velocity anisotropy (\code{$B}), rotational velocity (\code{$vrot}) and the
spin parameter (\code{$lambda}) are included in the output data frame. Finally, the dispersion
of particles off the plane of the disc is also given as \code{$sigma_z}, which describes the
standard deviation of particle positions about the z-axis.
}
\description{
The purpose of this function is to calculate the kinematic properties of a simulated galaxy,
specifically the velocity and dispersion of particles within certain user defined bins. The user
must specify which direction they wish to study the kinematics using \code{bin_type} (where
\code{= "r"} specifies out radial 3D spherical bins, \code{= "cr"} specifies radial 2D circular
bins, and \code{= "z"} specifies directly in 1D out of the plane of the galaxy). The default,
\code{bin_type = "r"} will return a comprehensive list of the simulation's kinematic properties
(i.e. contained radial mass and densities, velocities, velocity dispersions, anisotropy,
rotational and circular velocities and spin parameter). Other options for \code{bin_type} will
not contain the anisotropy, rotational and circular velocities or spin parameter as these are
only physical when defined using 3D spherical shells.
}
\examples{
 output = sim_analysis(filename   = system.file("extdata", 'S0_vignette', package="SimSpin"),
                       DM_profile = list("profile"="Hernquist", "DM_mass" = 184.9, "DM_a" = 34.5))

 output = sim_analysis(filename   = system.file("extdata", 'S0_vignette', package="SimSpin"),
                       ptype      = c(2,3),
                       bin_type   = "cr",
                       rmax       = 300,
                       rbin       = 100,
                       DM_profile = list("profile"="Hernquist", "DM_mass" = 184.9, "DM_a" = 34.5))

}
