% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_analysis.R
\name{sim_analysis}
\alias{sim_analysis}
\title{Kinematic analysis of simulation data.}
\usage{
sim_analysis(filename, bin_type = "r", rmax = 200, rbin = 200,
  ptype = NA)
}
\arguments{
\item{filename}{The Gadget file containing the particle information of the galaxy to be analysed.}

\item{bin_type}{The direction in which to bin the simulation model - "r" (default) bins radially in 3D spherical shells, "cr" bins radially in 2D circular rings,
"z" bins in 1D off the plane of the galaxy.}

\item{rmax}{The maximum radial coordinate considered within the simulated galaxy in kpc.}

\item{rbin}{The number of radial bins considered.}

\item{ptype}{The particle type/types to be extracted - NA (default) gives all particles in the simulation, 1 - gas, 2 - dark matter, 3 - disc, 4 - bulge, 5 - stars,
6 - boundary.}
}
\value{
A data frame containing kinematic features of radial bins including (at least) the outer radius of each bin (\code{$r}/\code{$cr}/\code{$z}), contained mass
(\code{$Mass}), logarithmic density (\code{$logp}), velocities and velocity dispersions (\code{$vr}/\code{$vcr}/\code{$vz} and \code{$sigma_vr}/\code{$sigma_vcr}/
\code{$sigma_vz}) and angular momentum magnitude (\code{$J}). In the case of \code{bin_dir = "r"} additionally the circular velocity (\code{$vc}), the velocity
anisotropy (\code{$B}), rotational velocity ($vrot) and the spin parameter (\code{$lambda}) are included in the output data frame.
}
\description{
The purpose of this function is to calculate the kinematic properties of a simulated galaxy, specifically the velocity and dispersion of particles within
certain user defined bins. The user must specify which direction they wish to study the kinematics using \code{bin_dir} (where \code{= "r"} specifies out
radially in 3D spherical bins, \code{= "cr"} specifies radially in 2D circular bins, and \code{= "z"} specifies directly in 1D out of the plane of the galaxy).
The default, \code{bin_dir = "r"} will return a comprehensive list of the simulation's kinematic properties (i.e. contained radial mass
and densities, velocities, velocity dispersions, anisotropy, rotational and circular velocities and spin parameter).
Other options for \code{bin_dir} will not contain the anisotropy, rotational and circular velocities or spin parameter as these are only physical when defined
using 3D spherical shells. If a simulation contains a vary large number of particles, it is possible to extract and analyse a sample of these to reduce
computational expense. This can be dome by specifying \code{samplerate = #a reduced number of particles}.
}
\examples{
\dontrun{
 sim_analysis(filename = "path/to/some/snapshot_XXX")

 sim_analysis(filename    = "path/to/some/snapshot_XXX",
              pype        = c(3,4),
              rmax        = 300,
              rbin        = 100,
              samplerate  = 1000000)
}

}
