% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Prepare simulation data for SimSpin}
\usage{
sim_data(filename, ptype = NA, SSP = FALSE, m2l_disc = 1,
  m2l_bulge = 1, m2l_star = 1, multi_thread = TRUE)
}
\arguments{
\item{filename}{The SimSpin HDF5 file containing the particle information of the galaxy.}

\item{ptype}{The particle type/types to be extracted - NA (default) gives all particles in the
simulation, 0 - gas, 1 - dark matter, 2 - disc, 3 - bulge, 4 - stars, 5 - boundary.}

\item{SSP}{Boolean specifying whether you wish the flux of the particle be calculated based
on the supplied Age/Metallicity information or scaled by the mass. Default is FALSE. If set
to TRUE and no Age/Metallicity information is present in the file, function will error.}

\item{m2l_disc}{The mass-to-light ratio of the disc component in solar units.}

\item{m2l_bulge}{The mass-to-light ratio of the bulge component in solar units.}

\item{m2l_star}{If no SSP file is specified and stellar particles exist in the file, the
mass-to-light ratio of the stellar component in solar units.}

\item{multi_thread}{A boolean specifying whether you would like to multi-thread the process.}
}
\value{
A list of data.frames containing the particle information (\code{$Part}) for each
particle type requested from the simulation. Each data frame contains the position (x, y, z)
and velocity (vx, vy, vz) information, along with the ID and mass of each particle. Also
associated with each PartType element in the list is an array of luminosities, \code{$Lum}. If
SSP information has been supplied, instead of Luminosities, a data frame \code{$SSP} will be
included in the list which contains the Metallicity and Ages of each particle such that spectra
can be constructed at later part in the process.
}
\description{
The purpose of this function is to take a SimSpin simulation file (an HDF5 file with particle
information subset in particle types, "PartTypeX") and output a data.frame that is easy for
processing throughout the rest of the package. This reduces the number of times that you need to
load the simulation into the package, which is the most time-consuming process. This also allows
you to combine the particle information with spectral information output from other codes.
}
\examples{
output = sim_data(system.file("extdata", 'SimSpin_example.hdf5', package="SimSpin"))
}
