% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope.R
\name{telescope}
\alias{telescope}
\title{A class to describe the basic properties of the observing telescope}
\usage{
telescope(
  type = "IFU",
  fov = 15,
  aperture_shape = "circular",
  wave_range = c(3700, 5700),
  spatial_res = 0.5,
  wave_res = 1.04,
  lsf_fwhm = 2.65
)
}
\arguments{
\item{type}{String that describes the type of telescope you wish to create.
Current pre-loaded types include "SAMI", "MaNGA", "CALIFA", and "Hector".
Input is NOT case sensitive. If you wish to specify different observing
properties below, set \code{type = "IFU"} to define your own telescope.}

\item{fov}{Numeric describing the field of view of the instrument in arcsec.}

\item{aperture_shape}{String to describe the shape of the IFU aperture.
Options include "circular", "hexagonal" or "square".}

\item{wave_range}{Numeric vector of length 2 describing the wave range of the
IFU (i.e. \code{c(wave_min, wave_max)}.}

\item{spatial_res}{Numeric describing the size of spatial pixels in arcsec.}

\item{wave_res}{Numeric describing the wavelength resolution in angstrom.}

\item{lsf_fwhm}{Numeric describing the full-width half-maximum of the Gaussian
line spread function.}
}
\value{
Returns an object of class "telescope" that describes the properties
of the instrument doing the observation. Required to run
\code{build_datacube()}.
}
\description{
The purpose of this function is to generate the observing telescope whose
properties will be used to compute the properties of the observation.
Several telescope types are included and can be accessed using the
\code{type} parameter. Alternatively, you may describe the properties of
another telescope by specifying \code{type = "IFU"} and listing the other
required properties explicitly.
}
\examples{
telescope = telescope(type="SAMI")

}
