% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_simspin_FITS.R
\name{write_simspin_FITS}
\alias{write_simspin_FITS}
\title{A function for making a mock spectral cube FITS file output}
\usage{
write_simspin_FITS(
  output_file,
  simspin_datacube,
  object_name,
  telescope_name,
  instrument_name,
  observer_name,
  input_simspin_file,
  mask = NA,
  galaxy_centre = c(0, 0, 0)
)
}
\arguments{
\item{output_file}{The path to the location where the FITS file should be
written.}

\item{simspin_datacube}{The list output from \code{build_datacube()}.}

\item{object_name}{A string that described the name of the observed object.}

\item{telescope_name}{A string that describes the name of the telescope used.}

\item{instrument_name}{A string that describes the used instrument on that
telescope.}

\item{observer_name}{A string that describes the name of the observer.}

\item{input_simspin_file}{A string describing the original SimSpin file
(i.e. the file output from make_simspin_file).}

\item{mask}{(Optional) A binary array describing the masked regions of the
cube/images.}

\item{galaxy_centre}{(Optional) A numeric array (x,y,z) describing the centre
of potential of the observed galaxy within it's simulation.}
}
\value{
Returns an .FITS file that contains a the generated data and
relevant header describing the mock observation.
}
\description{
The purpose of this function is to write the spectral cube generated using
the \code{build_datacube()} function to a FITS file.
}
\examples{
\dontrun{
ss_eagle = system.file("extdata", "SimSpin_example_EAGLE.hdf5", package = "SimSpin")
temp_loc = tempdir()
make_simspin_file(ss_eagle, output = paste(temp_loc, "spectra.Rdata", sep=""))
cube = build_datacube(simspin_file = paste(temp_loc, "spectra.Rdata", sep=""),
                     telescope = telescope(type="SAMI"),
                     observing_strategy = observing_strategy())
write_simspin_FITS(output_file = paste(temp_loc, "cube.FITS", sep=""),
                  simspin_datacube = cube, object_name = "SimSpin EAGLE example",
                  telescope_name = "AAO", instrument_name = "SAMI",
                  observer_name = "K.E.Harborne",
                  input_simspin_file = paste(temp_loc, "spectra.Rdata", sep=""))
unlink(paste(temp_loc, "spectra.fst", sep=""))
unlink(paste(temp_loc, "cube.FITS", sep=""))
}

}
